PAY-65: Add statistics at interface and trait + Adjust models to use the new trait
